requirements.txt
beautifultable==1.0.1
cachetools==5.0.0
DateTime==4.4
google-auth==2.6.0
google-auth-oauthlib==0.5.0
gspread==5.2.0
oauthlib==3.2.0
prettytable==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
requests-oauthlib==1.3.1
rsa==4.8
zope.interface==5.4.0
